pyproject.toml
[project]
name = "chain-peeper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dmitriy Strunin", email = "dstrunin@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.5.2",
    "ib-async>=2.1.0",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "streamlit>=1.57.0",
]

[project.scripts]
chain-peeper = "chain_peeper:main"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"
